feat: createCollection, deleteCollection and listCollections added
