+++ src/Data/DetectLabelsImagePropertiesSettingsData.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace MoeMizrak\Rekognition\Data;
 
+use MoeMizrak\Rekognition\Data\Contracts\RekognitionDataFormatInterface;
 use Spatie\LaravelData\Data;
 
 /**
@@ -12,7 +13,7 @@ use Spatie\LaravelData\Data;
  *
  * @class DetectLabelsImagePropertiesSettingsData
  */
-final class DetectLabelsImagePropertiesSettingsData extends Data
+final class DetectLabelsImagePropertiesSettingsData extends Data implements RekognitionDataFormatInterface
 {
     public function __construct(
         /*
@@ -25,9 +26,7 @@ final class DetectLabelsImagePropertiesSettingsData extends Data
     ) {}
 
     /**
-     * Convert the data to an AWS Rekognition array format.
-     *
-     * @return array
+     * @inheritDoc
      */
     public function toRekognitionDataFormat(): array
     {
