+++ src/Data/S3ObjectData.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace MoeMizrak\Rekognition\Data;
 
+use MoeMizrak\Rekognition\Data\Contracts\RekognitionDataFormatInterface;
 use Spatie\LaravelData\Data;
 
 /**
@@ -13,7 +14,7 @@ use Spatie\LaravelData\Data;
  *
  * @class S3ObjectData
  */
-final class S3ObjectData extends Data
+final class S3ObjectData extends Data implements RekognitionDataFormatInterface
 {
     public function __construct(
         /*
@@ -39,9 +40,7 @@ final class S3ObjectData extends Data
     ) {}
 
     /**
-     * Convert the data to an AWS Rekognition array format, excluding null values.
-     *
-     * @return array
+     * @inheritDoc
      */
     public function toRekognitionDataFormat(): array
     {
