+++ src/Data/SettingsData.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace MoeMizrak\Rekognition\Data;
 
+use MoeMizrak\Rekognition\Data\Contracts\RekognitionDataFormatInterface;
 use Spatie\LaravelData\Data;
 
 /**
@@ -13,7 +14,7 @@ use Spatie\LaravelData\Data;
  *
  * @class SettingsData
  */
-final class SettingsData extends Data
+final class SettingsData extends Data implements RekognitionDataFormatInterface
 {
     public function __construct(
         /*
@@ -32,9 +33,7 @@ final class SettingsData extends Data
     ) {}
 
     /**
-     * Convert the data to an AWS Rekognition array format, excluding null values.
-     *
-     * @return array
+     * @inheritDoc
      */
     public function toRekognitionDataFormat(): array
     {
