+++ src/RekognitionAPI.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace MoeMizrak\Rekognition;
 
 use Aws\Rekognition\RekognitionClient;
+use MoeMizrak\Rekognition\Helpers\RekognitionHelper;
 
 /**
  * RekognitionAPI abstract class responsible for encapsulating the RekognitionRequest class.
@@ -17,6 +18,10 @@ readonly abstract class RekognitionAPI
      * RekognitionAPI constructor.
      *
      * @param RekognitionClient $client
+     * @param RekognitionHelper $rekognitionHelper
      */
-    public function __construct(protected RekognitionClient $client) {}
+    public function __construct(
+        protected RekognitionClient $client,
+        protected RekognitionHelper $rekognitionHelper,
+    ) {}
 }
